backend/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
djongo==1.3.6
djangorestframework-simplejwt==5.3.0
pymongo==4.6.1
